+++ Core/Source/Renderer/AnimationSystem.cpp
@@ -9,7 +9,7 @@ void Mupfel::AnimationSystem::Update(double timestep) {
 	Mupfel::Registry& registry = Mupfel::Application::GetCurrentRegistry();
 	for (auto [e, animation] : registry.view<Animation>())
 	{
-		animation.elapsed += timestep;
+		animation.elapsed += static_cast<float>(timestep);
 		uint32_t step = static_cast<uint32_t>(animation.elapsed * animation.fps) % animation.frameCount;
 		animation.currentFrame = animation.firstFrame + step;
 	}
